Grabbing UID from Firebase for User, added OAuth & Apple
We now can grab the uid from the user and link that to SQL. I will link it tomorrow and perform a database call based on that UID to pull info. I also enabled OAuth (I'm still having issues with Google People API for some reason). Apple login code is there and should work, but I have to enable it on firebase to get functionality.