Shop Page Functionality
Currency now displays on the top right (I need to fix it so it updates on change but for now it's basically static until you reload the app)

You can now click on "Buy Skin" in the shop and it will deduct from your currency effectively buying the skin.

Functionality that still needs to happen, the skin needs to be part of an inventory or maybe it just auto swaps to your character? The store could be a kind of inventory? Not sure.

Added some new functions to flutterUser for this as well.